Major Revamp / Update
Firstly, we took into consideration. Was the help command, it should be more information and just be an overall better help command. It should be more helpful. Secondly we fixed some ticket glitches and made it slightly better. Then we revamped moderation commands to be a little bit better, as in making sure the user gets a DM. And adding all the fields. We also don't use pillow anymore for image manipulation.